# docs/swagger.yaml
basePath: /
definitions:
  categoryrequests.UpdateRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  entity.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  response.CategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/entity.Category'
        type: array
    type: object
  response.CategoryResponse:
    properties:
      category:
        $ref: '#/definitions/entity.Category'
    type: object
  response.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  response.IDResponse:
    properties:
      id:
        example: 123
        type: integer
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API for forum service
  title: Forum Service API
  version: "1.0"
paths:
  /categories:
    get:
      description: Retrieves a list of all categories.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all categories
          schema:
            $ref: '#/definitions/response.CategoriesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category. Requires admin role.
      parameters:
      - description: Category data to create. ID, CreatedAt, UpdatedAt will be ignored.
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/response.IDResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Deletes a category by its ID. Requires admin privileges.
      parameters:
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Category deleted successfully
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to delete category
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      description: Retrieves a specific category by its ID.
      parameters:
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved category
          schema:
            $ref: '#/definitions/response.CategoryResponse'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to get category
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a category by ID
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Updates a category's title and/or description by its ID. Requires
        admin privileges.
      parameters:
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Category update data
        in: body
        name: category_update
        required: true
        schema:
          $ref: '#/definitions/categoryrequests.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
        "400":
          description: Invalid category ID or request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update category
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a category
      tags:
      - categories
...
swagger: "2.0"